This zip file contains the full KROS project with LCD display
Note that the code for the LCD display is stored in the On Chip RAM used in the NIOS II Qsys system. Also a .pof file is generated so by simply download the .pof you will have the KROS function including the LCD display. The next step is to change the single bit per LED and replace it with a 3 bit value indicating the brightness of the LED. The LCD cannot perform brightness, so an OR of the 3 bit value per LED should be performed to indicate ON or OFF.